Towards a new flows API
Currently we register flows by creating a socket and binding it into the
flow we want. I believe it should be possible to register flows from the
application without affecting the LPG. This would allow the application
to have a single UDP listen (i.e., local port/ip) socket and just
register flows which are used to evaluate configurations via the cost
function.

Currently, there are two control functions multiplexed over
dnal_socket_bind:
    APPCTRL_SOCKET_UDPLISTEN,   // App -> Dragonet
    APPCTRL_SOCKET_UDPFLOW,     // App -> Dragonet

Effectively they do the same thing. The difference is that the first
specifies only the local port and local ip (similar to a traditional
socket bind), while the second allows more flexibility via specifying
the remote ip/port.

In this commit we merge both of them into a single command:
 - APPCTRL_SOCKET_UDPBIND

This leaves us to use APPCTRL_SOCKET_UDPFLOW for registering flows
without affecting the LPG. That is, allow the application to communicate
a set of flows to the stack without creating new sockets.