stack: remove ssPrevEndpoints
This was intended to be used as a way to figure out what changed in the
state when we were using endpoints for the state of stack. Using
ssPrevEndpoints, we could find the added/removed endpoints and implement
an incremental search.

With recent changes, we do not compute the cost function flows from
endpoints. Instead, the application registers them explicitly. Hence,
ssPrevEndpoints is no longer needed.